sharp: remove dependency and its own dependencies
We need to remove the sharp dependency as it gets pulled externally and
built inside the package.
The package then links to the `libvips-cpp.so.8.16.1` and it is detected
as a dependency that needs to be fulfilled by apk.
Since we are not able to honor the dependency, we remove it from the
dependency list from upstream so that we don't get into that situation.